+++ pages/sinclair.php
@@ -19,6 +19,7 @@ $template->assign_vars(array(
 	'GRAPH_TYPE_LC' => 'sinclair',
 	'GRAPH_TYPE' => 'Sinclair',
 	'MULTICHART' => false,
+	'MESSAGE' => 'The Sinclair Coefficient is a coefficient that can be used to estimate how much a weightlifter could lift if they were at the same level in the highest possible weight class.',
 	'RANGE' => $range
 	));
 $template->set_filenames(array(